LayoutTests/accessibility/mac/definition-list-term-expected.txt

term role
definition
definition
This tests that each of these ARIA grouping type roles map to the right role and subrole.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS obj.childAtIndex(0).role is 'AXRole: AXGroup'
PASS obj.childAtIndex(0).subrole is "AXSubrole: AXTerm"
PASS obj.childAtIndex(0).roleDescription is "AXRoleDescription: term"
PASS obj.childAtIndex(1).role is 'AXRole: AXGroup'
PASS obj.childAtIndex(1).subrole is "AXSubrole: AXDescription"
PASS obj.childAtIndex(1).roleDescription is "AXRoleDescription: description"
PASS obj.childAtIndex(1).role is 'AXRole: AXGroup'
PASS obj.childAtIndex(1).subrole is "AXSubrole: AXDescription"
PASS obj.childAtIndex(1).roleDescription is "AXRoleDescription: description"
PASS successfullyParsed is true

TEST COMPLETE
